+++ src/commands/index.ts
@@ -16,3 +16,4 @@ export * from "./RemoveProjectCommand";
 export * from "./CreateSolutionFolderCommand";
 export * from "./RemoveSolutionFolderCommand" ;
 export * from "./MoveToSolutionFolderCommand";
+export * from "./RenameSolutionItemCommand";