+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "vscode-solution-explorer" extension will be documented in this file.
 
-### 0.2.0
+### 0.2.1
 
 Refactor commands using command pattern.
 
@@ -17,6 +17,12 @@ Added add and remove packages commands in CPS projects.
 Added add and remove project reference commands in CPS projects.
 
 Added create project and solution commands.
+
+Added create and remove solution folder commands.
+
+Added move solution folder and project to solution folder commands.
+
+Added rename solution, solution folder and project commands.
 
 ## 0.0.9